+++ src/models/user.py
@@ -1,6 +1,6 @@
 import uuid
 
-from sqlalchemy import UUID, BigInteger, Integer
+from sqlalchemy import UUID, BigInteger
 from sqlalchemy.orm import Mapped, mapped_column
 
 from src.models.meta import Base
@@ -15,6 +15,6 @@ class User(Base):
         primary_key=True,
         default=uuid.uuid4,
     )
-    user_id: Mapped[int] = mapped_column(Integer, unique=True, index=True)
+    user_id: Mapped[int] = mapped_column(BigInteger, unique=True, index=True)
     subscription: Mapped[int] = mapped_column(BigInteger, nullable=True, default=None)
-    waiting_invoice: Mapped[int] = mapped_column(Integer, nullable=True, default=None)
+    waiting_invoice: Mapped[int] = mapped_column(BigInteger, nullable=True, default=None)
